22) Context API로 전역 값 관리 - 입력폼 자체에서 값 관리
